+++ middleman-core/fixtures/external-helpers/helpers/four_helper.rb
@@ -1,0 +1,3 @@
+module FourHelper
+  def four; "Four"; end
+end