+++ middleman-core/fixtures/external-helpers/helpers/three_helper.rb
@@ -1,0 +1,3 @@
+module ThreeHelper
+  def three; "Three"; end
+end