+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 * Don't re-minify files with ".min" in their name
 * Serve purely static folders directly (without source/ and config.rb)
 * Set ignored files and disable directory_indexes from YAML frontmatter
+* Automatically load helper modules in helpers/ directory
 
 2.0.14
 ====